Added needed mappings

See merge request logistics8371905/logistics!26

+++ backend/src/main/java/com/yukon/logistics/service/impl/TruckerServiceImpl.java
@@ -50,6 +50,11 @@ public class TruckerServiceImpl implements TruckerService {
 
     @Override
     public void deleteCityById(Long id) {
-        truckerRepository.deleteById(id);
+        if(truckerRepository.findById(id).isPresent()){
+            truckerRepository.deleteById(id);
+        }
+        else{
+            throw new EntityNotFoundException(id + "Trucker not found");
+        }
     }
 }
